LayoutTests/imported/w3c/web-platform-tests/css/css-grid/abspos/positioned-grid-descendants-007-expected.txt


FAIL .grid 1 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 1; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 1; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 2 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 2; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 2; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 3 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 4 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: auto; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 5 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: 2; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: 2; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 6 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 7 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 1; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 8 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 2; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 2; grid-row-end: 3; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 9 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 2; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 2; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
FAIL .grid 10 assert_equals: 
<div class="grid"><div style="grid-row-start: 1; grid-column-start: 1; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="45" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 3; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div><div style="grid-row-start: 2; grid-column-start: 2; grid-row-end: auto; grid-column-end: auto;">X<br>XX<div class="abspos" data-expected-width="0" data-expected-height="25" data-offset-x="25" data-offset-y="195" style="grid-column-start: auto; grid-column-end: 1; grid-row-start: 3; grid-row-end: auto; inset: auto 35px auto 25px;">XX</div></div></div>
width expected 0 but got -55
PASS .grid 11 
PASS .grid 12 
PASS .grid 13 
PASS .grid 14 
PASS .grid 15 
PASS .grid 16 
PASS .grid 17 
PASS .grid 18 
PASS .grid 19 
PASS .grid 20 
PASS .grid 21 
PASS .grid 22 
PASS .grid 23 
PASS .grid 24 
PASS .grid 25 
PASS .grid 26 
PASS .grid 27 
PASS .grid 28 
PASS .grid 29 
PASS .grid 30 
PASS .grid 31 
PASS .grid 32 
PASS .grid 33 
PASS .grid 34 
PASS .grid 35 
PASS .grid 36 
PASS .grid 37 
PASS .grid 38 
PASS .grid 39 
PASS .grid 40 
PASS .grid 41 
PASS .grid 42 
PASS .grid 43 
PASS .grid 44 
PASS .grid 45 
PASS .grid 46 
PASS .grid 47 
PASS .grid 48 
PASS .grid 49 
PASS .grid 50 
PASS .grid 51 
PASS .grid 52 
PASS .grid 53 
PASS .grid 54 
PASS .grid 55 
PASS .grid 56 
PASS .grid 57 
PASS .grid 58 
PASS .grid 59 
PASS .grid 60 
PASS .grid 61 
PASS .grid 62 
PASS .grid 63 
PASS .grid 64 
PASS .grid 65 
PASS .grid 66 
PASS .grid 67 
PASS .grid 68 
PASS .grid 69 
PASS .grid 70 
PASS .grid 71 
PASS .grid 72 
PASS .grid 73 
PASS .grid 74 
PASS .grid 75 
PASS .grid 76 
PASS .grid 77 
PASS .grid 78 
PASS .grid 79 
PASS .grid 80 
PASS .grid 81 
PASS .grid 82 
PASS .grid 83 
PASS .grid 84 
PASS .grid 85 
PASS .grid 86 
PASS .grid 87 
PASS .grid 88 
PASS .grid 89 
PASS .grid 90 
PASS .grid 91 
PASS .grid 92 
PASS .grid 93 
PASS .grid 94 
PASS .grid 95 
PASS .grid 96 
PASS .grid 97 
PASS .grid 98 
PASS .grid 99 
PASS .grid 100 
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
X
XXXX
